+++ classes/components/data/action/EditInterface.class.php
@@ -3,6 +3,5 @@ namespace Flex\Components\Data\Action;
 
 interface EditInterface{
     public function doEdit(?array $params=[]) : ?string;
-    public function doUpdate(?array $params=[]) : ?string;
 }
 ?>