+++ classes/components/data/action/EditUpdateInterface.class.php
@@ -1,0 +1,8 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+use Flex\Components\Data\Action\EditInterface;
+use Flex\Components\Data\Action\UpdateInterface;
+
+interface EditUpdateInterface extends EditInterface, UpdateInterface {}
+?>