+++ classes/components/data/action/InsertInterface.class.php
@@ -1,0 +1,7 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+interface InsertInterface{
+    public function doInsert(?array $params=[]) : ?string;
+}
+?>