+++ classes/components/data/action/PostInsertInterface.class.php
@@ -1,0 +1,9 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+use Flex\Components\Data\Action\PostInterface;
+use Flex\Components\Data\Action\InsertInterface;
+
+interface PostInsertInterface extends PostInterface, InsertInterface{
+}
+?>