+++ classes/components/data/action/PostInterface.class.php
@@ -3,6 +3,5 @@ namespace Flex\Components\Data\Action;
 
 interface PostInterface{
     public function doPost(?array $params=[]) : ?string;
-    public function doInsert(?array $params=[]) : ?string;
 }
 ?>