+++ classes/components/data/action/ReplInterface.class.php
@@ -1,0 +1,7 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+interface ReplInterface{
+    public function doRepl(?array $params=[]) : ?string;
+}
+?>