+++ classes/components/data/action/ReplyInterface.class copy.php
@@ -3,6 +3,5 @@ namespace Flex\Components\Data\Action;
 
 interface ReplyInterface{
     public function doReply(?array $params=[]) : ?string;
-    public function doRepl(?array $params=[]) : ?string;
 }
 ?>