+++ classes/components/data/action/ReplyReplInterface.class.php
@@ -1,0 +1,8 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+use Flex\Components\Data\Action\ReplyInterface;
+use Flex\Components\Data\Action\ReplInterface;
+
+interface ReplyReplInterface extends ReplyInterface,ReplInterface{}
+?>