+++ classes/components/data/action/UpdateInterface.class.php
@@ -1,0 +1,7 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+interface UpdateInterface{
+    public function doUpdate(?array $params=[]) : ?string;
+}
+?>