+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### - 2024-08-27
 - HashEncoder 와 Base64UrlEncoder 상속저의 해제
+- components/data/action 좀더 세세하게 interface 분리 및 결함
 
 ### - 2024-08-26
 - ArrayHelper ver1.3.1 버그 패치